+++ Cargo.toml
@@ -129,6 +129,7 @@ members = [
     "transaction/builder",
     "transaction/core",
     "transaction/core/test-utils",
+    "transaction/extra",
     "transaction/types",
     "util/b58-decoder",
     "util/build/enclave",
