+++ api/Cargo.toml
@@ -16,6 +16,7 @@ mc-crypto-multisig = { path = "../crypto/multisig" }
 mc-crypto-ring-signature-signer = { path = "../crypto/ring-signature/signer" }
 mc-transaction-builder = { path = "../transaction/builder" }
 mc-transaction-core = { path = "../transaction/core" }
+mc-transaction-extra = { path = "../transaction/extra" }
 mc-util-repr-bytes = { path = "../util/repr-bytes" }
 mc-util-serial = { path = "../util/serial" }
 mc-watcher-api = { path = "../watcher/api" }
