+++ api/src/convert/tx_out_gift_code.rs
@@ -4,10 +4,11 @@
 
 use crate::{external, printable, ConversionError};
 use mc_crypto_keys::{CompressedRistrettoPublic, RistrettoPrivate, RistrettoPublic};
+use mc_transaction_extra::TxOutGiftCode;
 
-/// Convert mc_transaction_core::TxOutGiftCode --> printable::TxOutGiftCode.
-impl From<&mc_transaction_core::TxOutGiftCode> for printable::TxOutGiftCode {
-    fn from(src: &mc_transaction_core::TxOutGiftCode) -> Self {
+/// Convert TxOutGiftCode --> printable::TxOutGiftCode.
+impl From<&TxOutGiftCode> for printable::TxOutGiftCode {
+    fn from(src: &TxOutGiftCode) -> Self {
         let mut tx_out_gift_code = printable::TxOutGiftCode::new();
         tx_out_gift_code.set_global_index(src.global_index);
         tx_out_gift_code
@@ -18,8 +19,8 @@ impl From<&mc_transaction_core::TxOutGiftCode> for printable::TxOutGiftCode {
     }
 }
 
-/// Convert from printable::TxOutGiftCode --> mc_transaction_core::TxOutGiftCode
-impl TryFrom<&printable::TxOutGiftCode> for mc_transaction_core::TxOutGiftCode {
+/// Convert from printable::TxOutGiftCode --> TxOutGiftCode
+impl TryFrom<&printable::TxOutGiftCode> for TxOutGiftCode {
     type Error = ConversionError;
 
     fn try_from(src: &printable::TxOutGiftCode) -> Result<Self, Self::Error> {
@@ -40,7 +41,6 @@ impl TryFrom<&printable::TxOutGiftCode> for mc_transaction_core::TxOutGiftCode {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use mc_transaction_core::TxOutGiftCode;
     use mc_util_from_random::FromRandom;
     use mc_util_serial::{decode, encode};
     use protobuf::Message;
@@ -48,11 +48,11 @@ mod tests {
     use rand_core::SeedableRng;
 
     #[test]
-    // test conversion between mc_transaction_core::TxOutGiftCode <-->
+    // test conversion between TxOutGiftCode <-->
     // printable::TxOutGiftCode.
     fn test_gift_code_serialization() {
         let mut rng: StdRng = SeedableRng::from_seed([1u8; 32]);
-        let source = mc_transaction_core::TxOutGiftCode::new(
+        let source = TxOutGiftCode::new(
             9001,
             RistrettoPrivate::from_random(&mut rng),
             RistrettoPublic::from_random(&mut rng),
@@ -61,7 +61,7 @@ mod tests {
         // Roundtrip from protobuf should return the same object
         {
             let external = printable::TxOutGiftCode::from(&source);
-            let recovered = mc_transaction_core::TxOutGiftCode::try_from(&external).unwrap();
+            let recovered = TxOutGiftCode::try_from(&external).unwrap();
             assert_eq!(source, recovered);
         }
