+++ api/tests/prost.rs
@@ -17,7 +17,8 @@ use mc_transaction_builder::{
     test_utils::get_input_credentials, EmptyMemoBuilder, ReservedSubaddresses,
     SignedContingentInputBuilder,
 };
-use mc_transaction_core::{Amount, SignedContingentInput};
+use mc_transaction_core::Amount;
+use mc_transaction_extra::SignedContingentInput;
 use mc_util_from_random::FromRandom;
 use mc_util_serial::round_trip_message;
 use mc_util_test_helper::{run_with_several_seeds, CryptoRng, RngCore};
