+++ fog/sample-paykit/src/cached_tx_data/memo_handler.rs
@@ -6,8 +6,8 @@ use displaydoc::Display;
 use mc_account_keys::{AccountKey, PublicAddress, ShortAddressHash, CHANGE_SUBADDRESS_INDEX};
 use mc_common::logger::{log, Logger};
 use mc_crypto_keys::{KeyError, RistrettoPublic};
-use mc_transaction_builder::{MemoDecodingError, MemoType};
 use mc_transaction_core::{get_tx_out_shared_secret, subaddress_matches_tx_out, tx::TxOut};
+use mc_transaction_extra::{MemoDecodingError, MemoType};
 use std::collections::HashMap;
 
 /// A handler object that holds a contacts list and tries to recieve and
