+++ test-vectors/memos/Cargo.toml
@@ -17,7 +17,7 @@ rand = { version = "0.8", default_features = false }
 mc-account-keys = { path = "../../account-keys" }
 mc-crypto-keys = { path = "../../crypto/keys", default-features = false }
 mc-test-vectors-definitions = { path = "../definitions" }
-mc-transaction-builder = { path = "../../transaction/builder" }
+mc-transaction-extra = { path = "../../transaction/extra" }
 mc-util-from-random = { path = "../../util/from-random" }
 mc-util-serial = { path = "../../util/serial", default-features = false }
 mc-util-test-vector = { path = "../../util/test-vector" }
