+++ transaction/core/src/domain_separators.rs
@@ -41,9 +41,6 @@ pub const TXOUT_MERKLE_NODE_DOMAIN_TAG: &str = "mc_tx_out_merkle_node";
 /// Domain separator for hashing the "nil" value in a Merkle tree.
 pub const TXOUT_MERKLE_NIL_DOMAIN_TAG: &str = "mc_tx_out_merkle_nil";
 
-/// Domain separator for hashing the confirmation number
-pub const TXOUT_CONFIRMATION_NUMBER_DOMAIN_TAG: &str = "mc_tx_out_confirmation_number";
-
 /// Domain separator for computing the extended message digest
 pub const EXTENDED_MESSAGE_DOMAIN_TAG: &str = "mc_extended_message";
