+++ transaction/core/src/lib.rs
@@ -20,10 +20,8 @@ mod domain_separators;
 mod input_rules;
 mod memo;
 mod revealed_tx_out;
-mod signed_contingent_input;
 mod token;
 mod tx_error;
-mod tx_out_gift_code;
 
 pub mod constants;
 pub mod encrypted_fog_hint;
@@ -42,13 +40,9 @@ pub use amount::{AmountError, MaskedAmount, MaskedAmountV1, MaskedAmountV2};
 pub use input_rules::{InputRuleError, InputRules};
 pub use memo::{EncryptedMemo, MemoError, MemoPayload};
 pub use revealed_tx_out::{try_reveal_amount, RevealedTxOut, RevealedTxOutError};
-pub use signed_contingent_input::{
-    SignedContingentInput, SignedContingentInputError, UnmaskedAmount,
-};
 pub use token::{tokens, Token};
 pub use tx::MemoContext;
 pub use tx_error::{NewMemoError, NewTxError, TxOutConversionError, ViewKeyMatchError};
-pub use tx_out_gift_code::TxOutGiftCode;
 
 // Re-export from transaction-types, and some from RingSignature crate.
 pub use mc_crypto_ring_signature::{Commitment, CompressedCommitment};
