+++ transaction/extra/Cargo.toml
@@ -1,11 +1,8 @@
 [package]
-name = "mc-transaction-builder"
+name = "mc-transaction-extra"
 version = "2.1.0-pre0"
 authors = ["MobileCoin"]
 edition = "2021"
-
-[features]
-test-only = []
 
 [dependencies]
 # External dependencies
@@ -22,14 +19,15 @@ zeroize = "1"
 
 # MobileCoin dependencies
 mc-account-keys = { path = "../../account-keys" }
+mc-crypto-digestible = { path = "../../crypto/digestible" }
 mc-crypto-hashes = { path = "../../crypto/hashes" }
 mc-crypto-keys = { path = "../../crypto/keys", default-features = false }
-mc-crypto-ring-signature-signer = { path = "../../crypto/ring-signature/signer", default-features = false }
-mc-fog-report-validation = { path = "../../fog/report/validation" }
+mc-crypto-ring-signature = { path = "../../crypto/ring-signature" }
+mc-crypto-ring-signature-signer = { path = "../../crypto/ring-signature/signer" }
 mc-transaction-core = { path = "../../transaction/core" }
-mc-transaction-extra = { path = "../../transaction/extra" }
 mc-transaction-types = { path = "../../transaction/types" }
 mc-util-from-random = { path = "../../util/from-random" }
+mc-util-repr-bytes = { path = "../../util/repr-bytes" }
 mc-util-serial = { path = "../../util/serial" }
 
 [target.'cfg(any(target_feature = "avx2", target_feature = "avx"))'.dependencies]
